move dask related functions in dedicated module
